# openapi.yaml
openapi: 3.0.3
info:
  title: Midas DVR / KLV API
  version: 1.0.0
  description: |
    HTTP API for source ingest, authoritative file workflows, HLS/KLV state,
    clips, snapshots, telemetry, and WebRTC signaling. File-backed operations
    use the private authoritative asset stored under mission-products/<productId>/source.
servers:
  - url: http://localhost:8090
    description: Default local server
...
paths:
...
  /sources/allocate:
    post:
      tags: [Sources]
      summary: Reserve a service-generated stream ID and future FMV product workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [missionId, sourceType]
              properties:
                missionId: { type: string, format: uuid }
                sourceType: { type: string, enum: [stream, file] }
      responses:
        '201':
          description: Internal stream and future product IDs allocated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceAllocation' }
...
  /sources/{streamId}/reset:
    post:
      tags: [Sources]
      summary: Validate a reserved source workspace before start
      description: Called before source start. It never removes product data.
      parameters: [ { $ref: '#/components/parameters/StreamId' } ]
      responses:
        '200':
          description: Stream start preflight passed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationResult' }
        '409': { $ref: '#/components/responses/Error' }
...
  /uploads/video/resumable:
    post:
      tags: [Uploads]
      summary: Create a resumable browser upload session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [streamId, filename, sizeBytes]
              properties:
                streamId: { type: string }
                filename: { type: string, example: mission.ts }
                sizeBytes: { type: integer, format: int64 }
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadSession' }
        '400': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
  /uploads/video/resumable/{streamId}/{uploadId}:
    head:
      tags: [Uploads]
      summary: Get the server-confirmed resumable upload offset
      parameters: [ { $ref: '#/components/parameters/StreamId' }, { $ref: '#/components/parameters/UploadId' } ]
      responses:
        '204':
          description: Offset returned in Upload-Offset response header
          headers:
            Upload-Offset: { schema: { type: integer, format: int64 } }
        '404': { description: Upload session not found }
    patch:
      tags: [Uploads]
      summary: Append one resumable upload chunk
      parameters:
        - { $ref: '#/components/parameters/StreamId' }
        - { $ref: '#/components/parameters/UploadId' }
        - name: Upload-Offset
          in: header
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/offset+octet-stream:
            schema: { type: string, format: binary }
      responses:
        '204':
          description: Chunk accepted; next offset is in Upload-Offset
          headers:
            Upload-Offset: { schema: { type: integer, format: int64 } }
        '409': { $ref: '#/components/responses/Error' }
  /uploads/video/resumable/{streamId}/{uploadId}/complete:
    post:
      tags: [Uploads]
      summary: Finalize a completed resumable upload
      parameters: [ { $ref: '#/components/parameters/StreamId' }, { $ref: '#/components/parameters/UploadId' } ]
      responses:
        '201':
          description: Authoritative asset created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetResult' }
        '409': { $ref: '#/components/responses/Error' }
...
  /uploads/video/local-copy:
    post:
      tags: [Uploads]
      summary: Copy a selected local-server video into a stream source folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [streamId, inputPath]
              properties:
                streamId: { type: string }
                inputPath: { type: string, description: Server path returned by local-files listing }
      responses:
        '201':
          description: Authoritative asset copied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetResult' }
        '400': { $ref: '#/components/responses/Error' }
  /probe/input:
    post:
      tags: [Sources]
      summary: Probe a stream URL with FFprobe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inputUrl]
              properties: { inputUrl: { type: string } }
      responses:
        '200': { description: Media capability probe }
...
components:
  parameters:
    StreamId:
      name: streamId
      in: path
      required: true
      schema: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$' }
    UploadId:
      name: uploadId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Request could not be completed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties: { ok: { type: boolean, example: false }, error: { type: string } }
    OperationResult:
      type: object
      properties:
        ok: { type: boolean }
        deletedEvents: { type: integer }
        deletedTargetLog:
          type: object
          properties:
            entries: { type: integer }
            fields: { type: integer }
    AssetResult:
      type: object
      properties:
        ok: { type: boolean }
        assetId: { type: string, example: 550e8400-e29b-41d4-a716-446655440000.ts }
        sizeBytes: { type: integer, format: int64 }
        sourceFilename: { type: string }
    UploadSession:
      type: object
      properties:
        ok: { type: boolean }
        uploadId: { type: string, format: uuid }
        uploadUrl: { type: string }
        offset: { type: integer, format: int64 }
...
    SourceAllocation:
      type: object
      required: [ok, streamId]
      properties:
        ok: { type: boolean }
        streamId: { type: string, description: Service-generated internal backing ID. }
        productId: { type: string, format: uuid, description: Reserved future FMV product ID. }
...